# backend/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "PwndocNG API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "/",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/data/languages": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "locale": {
                  "example": "any"
                },
                "language": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "locale": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                },
                "forEach(e": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/data/audit-types": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "templates": {
                  "example": "any"
                },
                "sections": {
                  "example": "any"
                },
                "hidden": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "templates": {
                  "example": "any"
                },
                "forEach(e": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/data/vulnerability-types": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "locale": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "locale": {
                  "example": "any"
                },
                "forEach(e": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/data/vulnerability-categories": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "sortValue": {
                  "example": "any"
                },
                "sortOrder": {
                  "example": "any"
                },
                "sortAuto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "forEach(e": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/data/sections": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "field": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "locale": {
                  "example": "any"
                },
                "text": {
                  "example": "any"
                },
                "icon": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "field": {
                  "example": "any"
                },
                "forEach(e": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/data/custom-fields": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fieldType": {
                  "example": "any"
                },
                "label": {
                  "example": "any"
                },
                "display": {
                  "example": "any"
                },
                "displaySub": {
                  "example": "any"
                },
                "size": {
                  "example": "any"
                },
                "offset": {
                  "example": "any"
                },
                "required": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "text": {
                  "example": "any"
                },
                "options": {
                  "example": "any"
                },
                "position": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "label": {
                  "example": "any"
                },
                "_id": {
                  "example": "any"
                },
                "display": {
                  "example": "any"
                },
                "fieldType": {
                  "example": "any"
                },
                "forEach(e": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/data/custom-fields/{fieldId}": {
      "delete": {
        "tags": [
          "Data"
        ],
        "description": "",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  }
}